+++ src/XRoadLib.Extensions.AspNetCore/XRoadLibMiddlewareExtensions.cs
@@ -1,18 +1,24 @@
 ﻿using System;
 using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Routing;
 
 namespace XRoadLib.Extensions.AspNetCore
 {
     public static class XRoadLibMiddlewareExtensions
     {
-        public static IApplicationBuilder UseXRoadLib(this IApplicationBuilder builder)
+        public static IApplicationBuilder UseXRoadLib(this IApplicationBuilder app, Action<IRouteBuilder> configureRoutes)
         {
-            return builder.UseXRoadLib(options => options);
-        }
+            if (app == null)
+                throw new ArgumentNullException(nameof(app));
 
-        public static IApplicationBuilder UseXRoadLib(this IApplicationBuilder builder, Func<XRoadLibOptions, XRoadLibOptions> optionBuilder)
-        {
-            return builder.UseMiddleware<XRoadLibMiddleware>(optionBuilder(new XRoadLibOptions()));
+            if (configureRoutes == null)
+                throw new ArgumentNullException(nameof(configureRoutes));
+
+            var routes = new RouteBuilder(app);
+
+            configureRoutes(routes);
+
+            return app.UseRouter(routes.Build());
         }
     }
 }